requirements.txt
websockets
pandas
ta
python-bot-telegram-bot